rename tool and go module
