+++ connector-sdk-0.13-migrator/go.mod
@@ -1,0 +1,5 @@
+module github.com/conduitio/tools/connector-sdk-0.13-migrator
+
+go 1.23.2
+
+require github.com/conduitio/yaml/v3 v3.3.0
