+++ connector-sdk-0.13-migrator/internal/common.go
