+++ connector-sdk-0.13-migrator/internal/delete_paramgen.go
