+++ connector-sdk-0.13-migrator/internal/delete_spec_go.go
