+++ connector-sdk-0.13-migrator/internal/goreleaser.go
