+++ connector-sdk-0.13-migrator/internal/scripts.go
