+++ connector-sdk-0.13-migrator/internal/tools_go.go
