+++ connector-sdk-0.13-migrator/internal/write_connector_yaml.go
