+++ connector-sdk-0.13-migrator/main.go
@@ -6,7 +6,7 @@ import (
 	"os"
 	"reflect"
 
-	"github.com/conduitio-labs/connector-migrator/internal"
+	"github.com/conduitio/tools/connector-sdk-0.13-migrator/internal"
 )
 
 var allMigrators = []internal.Migrator{
