--- specgen_connector_migrator/go.mod
@@ -1,5 +1,0 @@
-module github.com/conduitio-labs/connector-migrator
-
-go 1.23.2
-
-require github.com/conduitio/yaml/v3 v3.3.0
